requirements.txt
absl-py==0.9.0
argon2-cffi==20.1.0
astor==0.8.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.functools-lru-cache==1.6.1
bleach==3.2.2
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.12.5
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==2.9.2
cycler==0.10.0
dataclasses==0.8
datasets==1.2.1
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
entrypoints==0.3
findspark==1.4.2
flash==1.0.3
Flask==1.1.2
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
importlib-metadata==3.4.0
ipykernel==5.4.2
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Mako==1.1.4
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
multiprocess==0.70.11.1
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.8
pandas==1.1.5
pandocfilters==1.4.2
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
pip==20.1.1
prometheus-client==0.9.0
prompt-toolkit==3.0.14
py4j==0.10.9
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.7.4
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pywinpty==0.5.7
PyYAML==5.4.1
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.0
scikit-datasets==0.1.38
scikit-multilearn==0.2.0
scipy==1.4.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
tensorboard==2.1.1
tensorflow-cpu==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
Theano==1.0.5
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.49.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.34.2
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xxhash==2.0.0
yml==0.0.1
zipp==3.4.0
